third_party/blink/web_tests/platform/win/fast/forms/select-popup/popup-menu-nested-style-expected.txt

Check if an OPTION style is serialized with comparison with the parent OPTGROUP style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS params.children[1].type is "optgroup"
PASS params.children[1].children[0].type is undefined.
PASS params.children[1].children[0].style.color is "#000000"
PASS params.children[1].children[0].style.backgroundColor is "#ffffff"
PASS successfullyParsed is true

TEST COMPLETE

foo
bar
baz
